Refactor Layout Builder from Sidebar Menu
This change removes the Layout Builder option from the sidebar menu of the app. The component was commented out in SidebarMenuMain.js file as it was providing redundant functionalities, which were already being covered by other components in the sidebar. Aim is to improve navigability and user experience within the app.