Add new Products module and refactor Dashboard for products support
This commit adds new product module with various sub-components such as 'AddProduct', 'EditProduct', 'ShowProduct', 'Products', etc. The Products module is responsible for rendering product related user interface with provision for linking with the backend data. This commit also refactors the Dashboard for products support by renaming 'DashboardWrapper' to 'Dashboard' and adding ProductsRoutes to 'AccountRoutes'. The objective is to make our application more interactive and easy to navigate to improve the overall user experience.