Refactor user dashboard and add customer role
This commit accomplishes two main tasks: the separation of user dashboards based on role (customer or seller), and the creation of distinct routes for each user type. To begin, account/Dashboard was renamed to SellerDashboard, and new CustomerDashboard and SellerRoutes files were created for the routing of each role. Additionally, the user type check was refactored in AuthRoutes to accommodate these new user-specific dashboards. This update is pivotal to improve user experience by separating functionality and data between customers and sellers, hence better aligning with their specific tasks.