Change user type from customer to seller
Changed the value of the user type in the currentUser variable in the App.js file. The previous type was 'customer', but it was changed to 'seller' to reflect the specific functionality intended for this user type in the application. This user type change helps to maintain role-based access control in our app, ensuring users have access to appropriate actions based on their role. Finally, this will also assist with user testing from different role perspectives.