Add account types in Registration form
Introduced a dropdown in the Registration form for users to select their account type ("Seller" or "Customer"). This was necessary to differentiate the privileges and access rights of the two types of users. The new dropdown is implemented using select2 to maintain consistency with other form elements. The chosen account type is then mapped to its corresponding id.