Updated navigation and visual elements in various components
The navigation in ComingSoon, Error500 and Error404 components is switched from using 'Link' to 'useNavigate'. This was done to improve usability as it now enables user to return to the last screen instead of always redirecting to the dashboard. Also, it enhances user experience by asking the browser to navigate back in the history stack.

The visual elements in SellerDashboard are adjusted to improve readability. The size of icons, dynamic data display and the loading spinner are enhanced. The display of current user details in HeaderUserMenu is modified to fix a typo in displaying first_name twice and now it displays user's 'name'.