Remove forgot password feature and update user role display
Deleted the ForgotPassword component along with its corresponding path in AuthRoutes and removed linking from Login component as it was causing confusion for users and it was decided to handle password retrieval through a manual process. Also, updated how the user's role is displayed in the toolbar and user menu headers to provide clarity on user status. These changes are made to simplify the login process and better highlight the user's role in the system.