Refactor axios setup and network requests organization
Removed axiosCommonApi.js and extracted its functionality to apiService.js to create a setup for centralized API requests. This principle of having all requests in one file, which then exports the functions to use in our components, makes the management of requests easier and code cleaner. The functions from apiService.js are now used instead of direct axios requests in the components. Updated import statements for API_URL and moved apiUrl.js to the common directory for better organization.