Add UserBalanceCountUp and bid improvements
Introduced new component `UserBalanceCountUp` that provides live updates of user's balance. Modified bid system to provide better user feedback and realtime updates to product, bidding and user balances. Also, fixed the redirection bug in the product show page. Before, this component was only updating when the page was refreshed. Now, it provides live updates with the help of `useEffect`. Modified route and link in `ProductsRoutes` and `Products` component to provide more readable URLs. Removed `releaseProduct` function from `Products` component (seller-side) as it was not being used. Conditionally rendering the `bidAmount` and `highestBidAmountAnim` based on the product status. Included bid deposit and full payment options for users on the product page.