Remove countUpAnim and update README
Remove countUpAnim.update from both SellerDashboard.js and CustomerDashboard.js. These functions were causing issues in case of failure of API call and were unnecessary as the balance update is already handled in setDashboardData.

README.md updated with instructions on how to run the project with Docker and Docker Compose. Detailed steps from setting up the environment variables, installing dependencies, building the project to running the application are provided to offer a more comprehensive guide for new contributors or users.