Update metadata and description in index.html
Improved the description and added Open Graph metadata in index.html to help optimize social shares for the website. The Open Graph protocol allows any web page to become a rich object in a social graph. This will provide a better user experience as well as have potential SEO benefits. The updated description provides a clearer summary of the website's content.