Remove unused ref and adjust page styles
Removed an unused 'select2Ref' on EditProduct.js to prevent potential reference issues and code confusion. Replaced default logo with 'oktion.png' in HeaderWrapper.js to better represent the branding of the site. Added a dynamic color class for bidders' initials and adjusted element styling on ShowProduct.js to provide a more engaging design for bidders and improve user experience.